requirements.txt
backoff==2.2.1
comet==3.1.0
datasets==2.13.1
fairseq==0.12.2
mosestokenizer==1.2.1
msal==1.20.0
nevergrad==0.7.0
numpy==1.23.3
openai==1.74.0
PyYAML==6.0
Requests==2.31.0
rouge==1.0.1
sacrebleu==2.3.1
scikit_learn==1.1.2
termcolor==2.3.0
torch==2.0.0
tqdm==4.64.1
unbabel-comet
git+https://github.com/huggingface/transformers.git
jsonlines==3.1.0
git+https://github.com/feralvam/easse.git
accelerate==0.21.0
bitsandbytes==0.41.1
rich==13.6.0
python-dotenv==1.0.0
pydantic==2.5.2
streamlit==1.29.0